+++ tests/functional/lt_test_ire_pairing_key_slots.c
@@ -74,7 +74,7 @@ void lt_test_ire_pairing_key_slots(lt_handle_t *h)
     LT_LOG_INFO();
     for (uint8_t i = TR01_PAIRING_KEY_SLOT_INDEX_1; i <= TR01_PAIRING_KEY_SLOT_INDEX_3; i++) {
         LT_LOG_INFO("Reading pairing key slot %" PRIu8 " (should fail)...", i);
-        LT_TEST_ASSERT(LT_L3_PAIRING_KEY_EMPTY, lt_pairing_key_read(h, read_key, i));
+        LT_TEST_ASSERT(LT_L3_SLOT_EMPTY, lt_pairing_key_read(h, read_key, i));
         LT_LOG_INFO();
     }
     LT_LOG_LINE();
@@ -143,7 +143,7 @@ void lt_test_ire_pairing_key_slots(lt_handle_t *h)
         LT_TEST_ASSERT(LT_OK, lt_pairing_key_invalidate(h, i));
 
         LT_LOG_INFO("Reading pairing key slot %" PRIu8 " (should fail)...", i);
-        LT_TEST_ASSERT(LT_L3_PAIRING_KEY_INVALID, lt_pairing_key_read(h, read_key, i));
+        LT_TEST_ASSERT(LT_L3_SLOT_INVALID, lt_pairing_key_read(h, read_key, i));
 
         LT_LOG_INFO("Writing to pairing key slot %" PRIu8 " (should fail)...", i);
         LT_TEST_ASSERT(LT_L3_FAIL, lt_pairing_key_write(h, zeros, i));