+++ CHANGELOG.md
@@ -28,6 +28,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed `lt_dev_stm32_nucleo_f439zi` to `lt_dev_stm32_nucleo_f439zi_t`.
 - Added sections **Supported Host Platforms** and **Supported Cryptographic Functionality Providers** into **Other** section in the Libtropic documentation.
 - Added dependencies on [micro-ecc](https://github.com/kmackay/micro-ecc) and [ed25519](https://github.com/orlp/ed25519/tree/master) repositories in functional tests (for verification of signatures calculated by TROPIC01).
+- Updated L3 result values according to Application FW 1.1.0:
+  - Renaming inside `lt_l3_process.h` (not exposed to users directly).
+  - Renaming of `lt_ret_t` values:
+    - `LT_L3_PAIRING_KEY_EMPTY` to `LT_L3_SLOT_EMPTY`.
+    - `LT_L3_PAIRING_KEY_INVALID` to `LT_L3_SLOT_INVALID`.
+    - `LT_L3_ECC_INVALID_KEY` to `LT_L3_INVALID_KEY`.
+    - `LT_L3_R_MEM_DATA_WRITE_WRITE_FAIL` to `LT_L3_SLOT_NOT_EMPTY`.
+    - `LT_L3_R_MEM_DATA_WRITE_SLOT_EXPIRED` to `LT_L3_SLOT_EXPIRED`.
+    - `LT_L3_MCOUNTER_UPDATE_UPDATE_ERR` to `LT_L3_UPDATE_ERR`.
+- Order of values inside `lt_ret_t`.
+- Renamed `LT_L2_STATUS_NOT_RECOGNIZED` to `LT_L2_STATUS_UNKNOWN`.
+- `lt_l3_decrypt_response()` returns `LT_L3_RESULT_UNKNOWN` instead of `LT_FAIL` if the L3 result value is unknown.
+
 
 ### Added
 - Possibility to measure test coverage with the TROPIC01 model.
@@ -44,6 +57,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - In HALs that don't support it, raise compilation error.
   - Implement support in Unix SPI HAL.
 - Support for MbedTLS v4 crypto backend.
+- `enum lt_ret_t`:
+  - `LT_L3_RESULT_UNKNOWN` for unknown L3 result values from TROPIC01.
+  - `LT_L3_HARDWARE_FAIL` to reflect the new L3 result value from TROPIC01.
 
 ### Fixed
 - `lt_ex_show_chip_id_and_fwver`: reboot back to Application mode in the end.